More work on siap, particularly on the data model.
Tuples now serialize to arrays (need to check if that's a problem for
querulator, but probably not, it uses lists for SQL sets).

A test data set now ingests with what I believe are all parameters necessary
of a complete siap answer.

In consequence, we now have SQL array types to worry about.  typesystems
and the (still horribly named) parsing.typeconversions now handle them
more or less.