Fixes to the upload service.
Added a service for "machine" uploads with signalling over multipart/form-data,
http response codes and text messages.

Factored custom error messaging out of siapservice to re-use it in machine
upload service.

Added a utils directory for more valuable hacks (like, in this case, the
upload program that talks to the machine upload service).