Built static delivery of files from input directories via a special .static. sentinel and a new staticData attribute on services.
More work on registry responses.

There's a new standard core that always returns a file (but that's
really nonsense, because you'll probably want to use custom templates
when you think you want such a thing).

gavoimp now checks for the presence of certain meta keys necessary for
valid registry records and skips records that do not have them.  That's
a bit silly for local records, but it's probably a good idea to set
these things anyway.