Lots of fixes to registry code to make it play with the Registry of Registries.
queryServicesList now takes another argument, source, specifying the
data id in services.vord to use for querying.  This is done so resources
without interfaces and/or sets can be found.

Another bad hack to field copying: RowsetGrammar will leave a field's
source alone if copying from another RowsetGrammar.  We need to come up
with a *good* way of inheriting, badly.