Added a load test (reusing the regression tests)
Hacked in wcscon to have sane FK5 to FK4 transform.  Not sure if that's
going to be the way to add all those horrible transforms, but I need
that for the crappy USNO-B mess.

hourangleToDeg and dmsToDeg now understand "compressed" format with
sepChar=""

Fixed errors that would cause the DbOptions widget to crash when sorting
or limiting were seperately disabled.

Removed querulator wrapper qurun together with toplevel "web" directory.

Salvaged useful stuff out of (querulator-related) web_static and removed
it.