New module gavo.nodebuilder containing the utils.NodeBuilder.  It got too large for utils, and while we're refactoring XML stuff...
New module xmlstan for the stan-like XML building stuff so far only used
in the registry.

Revamped WDSL module to use xmlstan and return descriptions for SOAP
services.

New soaprenderer for handling SOAP communication.

A couple of last-minute fixes for metadata in the registry.  Finally
decided to drop the silly curation. prefix.