Started an implementation for STC-X and STC-S parsing and generation; this isn't anywhere near useful yet.
Moved table preparation for updating import to DataSet.exportToSql again
(to make it work:-); the owningCondition on shared tables is removed for
updating imports since we want to keep shared records already in the
tables.

"raw" data is now unsignedByte/* in VOTables and the registry.  This
probably won't quite do, since we shouldn't expose raw at all.

Fixed DC.publisher for registry to use the correct namespace.