Additional work to make tableinfos more useful.
MetaMixins now know how to deep copy themselves.  This is now used when
doing the original="..." hack in RDs; this ensures that meta info
on tables stays sane.

New -z option to gavoimp to dump you in a debugger on errors.

Made Positions and Q3CPositions base on a new RD __system__/scs.vord