Refactored the whole RD macro mess into a new module gavo.macros
New StandardMacroMixin for things with rds in gavo.macros.

MacroPackages now have a getExpander method returning an expander for
them and caching it.

Classes mixing in MetaMixin also mixing in MacroPackage now get
macro expansion when their stuff is rendered through HTMLMetaBuilder.
Maybe we want to generalize that mechanism.

Moved makeSitePath to macros.  It's a pain, but I guess some more of
these "simple" functions will have to go there over time.  Or do we want
a module utils?  Of course, we already have one:-)