Pulled macros into resdef
New module resdef.common.

New module resdesc for the resource descriptor.

New attribute def resdef.common.RDAttribute for automatic location
of the resource descriptor in its children.

Moved old DataFieldList to new common.ColumnList and wrote an
AttributeDef bridging it to Structures.

Changed structure parsing to only add structures when they are
completely parsed, including callback processing.  Therefore, the
misguided createOn method on StructAttributes was replaced by create.

Clarified nomenclature in attrdef.  There is a parent class (that embeds
the attribute) and instance (which is an instance of the parent class);
The latter was formerly called parent as well, which was confusing.
There's still some unfortunate nomenclature about in that in the module,
we talk about parent classes and structure parents, but I guess that's
tolerable.

structures now check for the presence of an attribute before setting
a default; this allows read-only properties as attributes.

structures can now define an elementCompleted callback that is called
when parsing from XML has finished the element and no new information
will come in (through this channel).  It feels a bit fragile in the view
of the possibility of manually constructed structures, but I get a
feeling that manual construction of structures is unsexy anyway.

Moved ScriptMacroTest out of misctest into a new module macrotest
and adapted it to the new structure.