Much work on rmkdef to make it work and give sensible error messages.
Defaults are now supported in RowmakerDefs.

Basic rowmapping works.

Added dddef and rmkdef (forgot about them in the last commit)

Structure parents are no longer weak references.  For one, we probably
don't want them to go away frequently, so the circular references don't
hurt.  More importantly, with deferred addition, there are cases in
which the parent reference from the currently parsed item is the only
reference we have, with bad consequences if it's weak.

New DictAttribute (for now in rmkdef, but should go to base.attrdef when
someone else needs it)