Made metatable work by splitting it into a MetaTableMixin to DBTable and the old MetaTableHandler for read access.
RDs now have a sourceId attribute again.

base.xmlstruct.parseFromStream now accepts instanciated structures as
roots of the future tree.

You can now pass nometa to dbtable to suppress entering the table
to the dc_tables (I guess that's only interesting for programmatic
access; otherwise, we'd have to think about an attribute on tabledef for
that).