Refactored rowmaker procs and frieds again.  Now they do "late" compiling (i.e., not in onElementComplete but rather in getDefinition, which gets called by rowmaker.compile.  This is very helpful with predefined procedures.
Scrapped copyableAttribute since it clashes with inheritance.  Instead,
AttributeDefinitions now have an attribute copyable for the same
purpose.

Defaults on procedure arguments are now evaluated sensibly; they
generate base.Undefined defaults in the rowmaker locals which are
then filled out in a prologue to the proc code.  This ensures macro
expansions in them are executed in the same context as macro expansions
in the code body.

To make things work more smoothly with row generators, RowIterators
should now override the _iterRows rather than __iter__.
RowIterator.__iter__ now runs the row generators if necessary.

Gave RDFuncs attributes for "computations of constants".  This will be
used in instanciating procs from predefined procedures, but isn't yet.