Made the internal error message a bit more urging.
The dc_tables tables are now called columnmeta instead of
fielddescriptions and tablemeta instead of dc_tables, and they now live
in a schema "dc".  This helps parallel operation of the old and the new
software, and the names are better, too.

Everything in the dc schema is now ignored by dbtable meta creation code
to alleviate pains when bootstrapping.

options fromDB is now an ActionAttribute; this means we actually
evaluate noQueries.