Input field selection can now be influenced by the renderer, i.e., service.getInputFields is now only called by a classmethod on grend.ServiceBasedRenderer.
You can now set properties onlyForRenderer and notForRenderer on
inputKeys and they will be interpreted in the way you might expect.
This required to let renderers have a say in the service's selection
of inputDDs, and this in turn the whole interface between renderers
and services.  Ah, well.

CondDescs will now check if their input parameters are present in the
input row and make themselves invisible when that is not the case.