taprunner no longer kills and exits from within a signal handler.
tap processQueue is now protected by a lock (which shouldn't be strictly
necessary, but it saves effort and reduces potential pitfalls, so it's
probably a good thing anyway).

The TAP dead job reaper is now only executing if the state transition
is in the server process, and it doesn't push jobs into ERROR just
because it cannot see their pids (that needs further tought).

Fixed a typo in Changes.