SP_Ace: here I add the log for the previous version (r4565) which, for unknown reasons, was not recorded.
In general, version r4565 differ from the previous one for accounting for 
possible non constant dispersion over the spectrum. 
The arrays rad_pix, norm_rad, and disp have dimension equal to the number of pixels and account 
for the local dipersion pixel by pixel.
The changed files were:
space_5.4.0.f95
interfaces.f95
make_model.f95
data_lib.f95 (here the line 8750.476 is included as line to reject)
utils.f95
share.f95
read_sp_ll.f95
fit_cont.f95

--------------------------
For the present version, the change is only the removing one line in utils.f95 to allow SP_Ace 
to plot lines that have been neglected. The pixels of such lines are anyway neglected because 
they have weight=0, but the lines are plotted in any case. This is because in some spectra 
(in particular for low resolution spectra) this improve the fit of blended features in which
lie close or over neglected intervals.